+++ tests/test_kit/mocks.py
@@ -406,8 +406,6 @@ class MockExecutionClient(ExecutionClient):
     ----------
     client_id : ClientId
         The client ID.
-    account_id : AccountId
-        The account_id for the client.
     account_type : AccountType
         The account type for the client.
     base_currency : Currency, optional
@@ -425,7 +423,6 @@ class MockExecutionClient(ExecutionClient):
     def __init__(
         self,
         client_id,
-        account_id,
         account_type,
         base_currency,
         msgbus,
@@ -436,7 +433,6 @@ class MockExecutionClient(ExecutionClient):
     ):
         super().__init__(
             client_id=client_id,
-            account_id=account_id,
             account_type=account_type,
             base_currency=base_currency,
             msgbus=msgbus,
@@ -496,8 +492,6 @@ class MockLiveExecutionClient(LiveExecutionClient):
     ----------
     client_id : ClientId
         The client ID.
-    account_id : AccountId
-        The account_id for the client.
     account_type : AccountType
         The account type for the client.
     base_currency : Currency, optional
@@ -518,7 +512,6 @@ class MockLiveExecutionClient(LiveExecutionClient):
         self,
         loop,
         client_id,
-        account_id,
         account_type,
         base_currency,
         instrument_provider,
@@ -530,7 +523,6 @@ class MockLiveExecutionClient(LiveExecutionClient):
         super().__init__(
             loop=loop,
             client_id=client_id,
-            account_id=account_id,
             account_type=account_type,
             base_currency=base_currency,
             instrument_provider=instrument_provider,
@@ -540,6 +532,7 @@ class MockLiveExecutionClient(LiveExecutionClient):
             logger=logger,
         )
 
+        self._set_account_id(AccountId(client_id.value, "001"))
         self._order_status_reports: Dict[VenueOrderId, OrderStatusReport] = {}
         self._trades_lists: Dict[VenueOrderId, list[ExecutionReport]] = {}