+++ tests/unit_tests/backtest/test_backtest_exchange_l2_mbp.py
@@ -64,7 +64,6 @@ class TestL2OrderBookExchange:
         )
 
         self.trader_id = TestStubs.trader_id()
-        self.account_id = TestStubs.account_id()
 
         self.msgbus = MessageBus(
             trader_id=self.trader_id,
@@ -124,7 +123,6 @@ class TestL2OrderBookExchange:
 
         self.exec_client = BacktestExecClient(
             exchange=self.exchange,
-            account_id=self.account_id,
             msgbus=self.msgbus,
             cache=self.cache,
             clock=self.clock,