+++ tests/unit_tests/live/test_live_execution_client.py
@@ -112,7 +112,6 @@ class TestLiveExecutionClient:
         self.client = MockLiveExecutionClient(
             loop=self.loop,
             client_id=ClientId(SIM.value),
-            account_id=TestStubs.account_id(),
             account_type=AccountType.CASH,
             base_currency=USD,
             instrument_provider=InstrumentProvider(),