+++ tests/unit_tests/trading/test_trading_strategy.py
@@ -74,7 +74,6 @@ class TestTradingStrategy:
         )
 
         self.trader_id = TestStubs.trader_id()
-        self.account_id = TestStubs.account_id()
 
         self.msgbus = MessageBus(
             trader_id=self.trader_id,
@@ -141,7 +140,6 @@ class TestTradingStrategy:
 
         self.exec_client = BacktestExecClient(
             exchange=self.exchange,
-            account_id=self.account_id,
             msgbus=self.msgbus,
             cache=self.cache,
             clock=self.clock,